Move provisioning debug assertions into provisioner tests
Inject errors with known bodies and hardcoded stacks via __setErrors
so the provisioner-actions tests cover debug output directly. Delete
the dedicated provisioning-explainer test file.
